chore:clena up the workflow graph demodata by moving it to utils
